# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Timetracker API Server"
        }
    ],
    "paths": {
...
        "/assignments": {
            "get": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Get list of assignments",
                "description": "Returns list of assignments",
                "operationId": "getAssignmentsList",
                "parameters": [
                    {
                        "name": "Page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Per page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "work_description": {
                                                        "type": "string"
                                                    },
                                                    "developer_description": {
                                                        "type": "string"
                                                    },
                                                    "id_clients": {
                                                        "type": "integer"
                                                    },
                                                    "id_statuses": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "work_description": "test",
                                                    "developer_description": "primer",
                                                    "id_clients": 1,
                                                    "id_statuses": 2,
                                                    "created_at": "2021-04-23T14:27:57.000000Z",
                                                    "updated_at": "2021-04-23T14:27:57.000000Z",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 2,
                                                    "work_description": "novi",
                                                    "developer_description": "primer",
                                                    "id_clients": 2,
                                                    "id_statuses": 2,
                                                    "created_at": "2021-04-23T14:31:20.000000Z",
                                                    "updated_at": "2021-04-23T14:31:20.000000Z",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 3,
                                                    "work_description": "noviProjekt",
                                                    "developer_description": "primer",
                                                    "id_clients": 2,
                                                    "id_statuses": 2,
                                                    "created_at": "2021-04-23T14:31:31.000000Z",
                                                    "updated_at": "2021-04-23T14:31:31.000000Z",
                                                    "deleted_at": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Store new assignment",
                "description": "Returns assignment data",
                "operationId": "storeAssignment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "work_description": {
                                        "type": "string"
                                    },
                                    "developer_description": {
                                        "type": "string"
                                    },
                                    "id_clients": {
                                        "type": "integer"
                                    },
                                    "id_statuses": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "work_description": {
                                            "type": "string"
                                        },
                                        "developer_description": {
                                            "type": "string"
                                        },
                                        "id_clients": {
                                            "type": "integer"
                                        },
                                        "id_statuses": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "work_description": "test work description",
                                        "developer_description": "test developer description",
                                        "id_clients": 1,
                                        "id_statuses": 3,
                                        "updated_at": "2021-04-28T17:39:37.000000Z",
                                        "created_at": "2021-04-28T17:39:37.000000Z",
                                        "id": 5
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/assignmentsClient/{id}": {
            "get": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Get list of assignments for client",
                "description": "Returns list of assignments for specific client",
                "operationId": "getAssignmentsList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "work_description": {
                                                        "type": "string"
                                                    },
                                                    "developer_description": {
                                                        "type": "string"
                                                    },
                                                    "id_clients": {
                                                        "type": "integer"
                                                    },
                                                    "id_statuses": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "work_description": "test",
                                                    "developer_description": "primer",
                                                    "id_clients": 1,
                                                    "id_statuses": 2,
                                                    "created_at": "2021-04-23 14:27:57",
                                                    "updated_at": "2021-04-23 14:27:57",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 4,
                                                    "work_description": "stari",
                                                    "developer_description": "primer",
                                                    "id_clients": 1,
                                                    "id_statuses": 1,
                                                    "created_at": "2021-04-23 14:31:42",
                                                    "updated_at": "2021-04-23 14:31:42",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 5,
                                                    "work_description": "test work description",
                                                    "developer_description": "test developer description",
                                                    "id_clients": 1,
                                                    "id_statuses": 3,
                                                    "created_at": "2021-04-28 17:39:37",
                                                    "updated_at": "2021-04-28 17:39:37",
                                                    "deleted_at": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}